feat(cm-account): implement Messenger Fee and Balances tab updates
