+++ src/views/partners/MyMessenger.tsx
@@ -1,32 +1,12 @@
-import { ContentCopy, RefreshOutlined } from '@mui/icons-material'
-import {
-    Box,
-    Button,
-    Checkbox,
-    CircularProgress,
-    DialogContent,
-    DialogTitle,
-    Divider,
-    FormControlLabel,
-    IconButton,
-    Link,
-    OutlinedInput,
-    TextField,
-    Typography,
-} from '@mui/material'
-import React, { useCallback, useEffect, useMemo, useState } from 'react'
+import { ContentCopy } from '@mui/icons-material'
+import { Box, Link, TextField, Typography } from '@mui/material'
+import React, { useEffect, useState } from 'react'
 import { useAppDispatch, useAppSelector } from '../../hooks/reduxHooks'
 import { fetchBusinessFields, fetchPartners } from '../../redux/slices/partnersSlice/utils'
 
-import { mdiClose } from '@mdi/js'
-import Icon from '@mdi/react'
-import { ethers } from 'ethers'
+import Blockies from 'react-blockies'
 import { useNavigate } from 'react-router'
-import store from 'wallet/store'
-import Alert from '../../components/Alert'
-import DialogAnimate from '../../components/Animate/DialogAnimate'
 import MainButton from '../../components/MainButton'
-import { ERC20_BALANCE_ABI } from '../../constants/apps-consts'
 import { usePartnerConfigurationContext } from '../../helpers/partnerConfigurationContext'
 import { usePartnerConfig } from '../../helpers/usePartnerConfig'
 import { useSmartContract } from '../../helpers/useSmartContract'
@@ -36,320 +16,12 @@ import { updateNotificationStatus } from '../../redux/slices/app-config'
 import { getActiveNetwork } from '../../redux/slices/network'
 import { Configuration } from './Configuration'
 
-const AmountInput = ({ amount, onAmountChange, onMaxAmountClick, maxAmount }) => {
-    const handleChange = e => {
-        const newAmount = e.target.value
-        if (newAmount === '' || /^\d*\.?\d*$/.test(newAmount)) {
-            onAmountChange(newAmount)
-        }
-    }
-
-    return (
-        <Box sx={{ display: 'flex', flexDirection: 'column', gap: '8px', width: '100%' }}>
-            <Box sx={{ display: 'flex', alignItems: 'center', gap: '12px', width: '100%' }}>
-                <Typography
-                    sx={{
-                        flex: '0 0 15%',
-                        padding: '6px 12px',
-                        gap: '6px',
-                        borderRadius: '8px',
-                        border: '1px solid #475569',
-                        backgroundColor: theme =>
-                            theme.palette.mode === 'dark' ? '#0F182A' : '#F1F5F9',
-                        borderWidth: '1px',
-                    }}
-                    variant="caption"
-                >
-                    Amount
-                </Typography>
-                <OutlinedInput
-                    value={amount}
-                    onChange={handleChange}
-                    inputProps={{
-                        inputMode: 'decimal',
-                        pattern: '[0-9]*',
-                    }}
-                    sx={theme => ({
-                        width: '100%',
-                        height: '40px',
-                        p: '8px 16px',
-                        border: `solid 1px ${theme.palette.card.border}`,
-                        borderRadius: '12px',
-                        fontSize: '14px',
-                        lineHeight: '24px',
-                        fontWeight: 500,
-                        '.MuiOutlinedInput-notchedOutline': {
-                            border: 'none',
-                        },
-                    })}
-                />
-                <Button
-                    variant="contained"
-                    onClick={onMaxAmountClick}
-                    sx={{
-                        flex: '0 0 16%',
-                        padding: '6px 12px',
-                        gap: '6px',
-                        borderRadius: '8px',
-                        border: '1px solid #475569',
-                        borderWidth: '1px',
-                        '&:hover': {
-                            borderWidth: '1px',
-                            boxShadow: 'none',
-                        },
-                    }}
-                >
-                    <Typography variant="caption">Max</Typography>
-                </Button>
-            </Box>
-            <Typography variant="caption" sx={{ alignSelf: 'flex-end' }}>
-                Max available: {maxAmount} CAM
-            </Typography>
-        </Box>
-    )
-}
-
-const AddressInput = ({ address, onAddressChange, onMyAddressClick }) => {
-    return (
-        <Box sx={{ display: 'flex', alignItems: 'center', gap: '12px', width: '100%' }}>
-            <Typography
-                sx={{
-                    padding: '6px 12px',
-                    gap: '6px',
-                    borderRadius: '8px',
-                    border: '1px solid #475569',
-                    backgroundColor: theme =>
-                        theme.palette.mode === 'dark' ? '#0F182A' : '#F1F5F9',
-                    borderWidth: '1px',
-                    flex: '0 0 15%',
-                }}
-                variant="caption"
-            >
-                To Address
-            </Typography>
-            <OutlinedInput
-                value={address}
-                onChange={e => onAddressChange(e.target.value)}
-                sx={theme => ({
-                    width: '100%',
-                    height: '40px',
-                    p: '8px 16px',
-                    border: `solid 1px ${theme.palette.card.border}`,
-                    borderRadius: '12px',
-                    fontSize: '14px',
-                    lineHeight: '24px',
-                    fontWeight: 500,
-                    '.MuiOutlinedInput-notchedOutline': {
-                        border: 'none',
-                    },
-                })}
-            />
-            <Button
-                variant="contained"
-                onClick={onMyAddressClick}
-                sx={{
-                    padding: '6px 12px',
-                    gap: '6px',
-                    borderRadius: '8px',
-                    border: '1px solid #475569',
-                    borderWidth: '1px',
-                    '&:hover': {
-                        borderWidth: '1px',
-                        boxShadow: 'none',
-                    },
-                    flex: '0 0 16%',
-                }}
-            >
-                <Typography variant="caption">My Address</Typography>
-            </Button>
-        </Box>
-    )
-}
-
-const CamWithdraw = ({ setOpen, token, fetchTokenBalances }) => {
-    const { wallet, contractCMAccountAddress } = useSmartContract()
-    const [address, setAddress] = useState('')
-    const [amount, setAmount] = useState('')
-    const [confirm, setConfirm] = useState(false)
-    const [isValidAddress, setIsValidAddress] = useState(false)
-    const [loading, setLoading] = useState(false)
-    const [amountError, setAmountError] = useState('')
-    const { withDraw, transferERC20 } = usePartnerConfig()
-    const { getBalanceOfAnAddress, balanceOfAnAddress: balance } = useWalletBalance()
-
-    const maxAmount = useMemo(() => {
-        if (token) return parseFloat(token.balance)
-        const balanceParsed = parseFloat(balance)
-        if (isNaN(balanceParsed)) {
-            return '0.00'
-        }
-        return Math.max(balanceParsed - 100, 0).toFixed(2)
-    }, [balance, token])
-
-    const handleAddressChange = useCallback(newAddress => {
-        setAddress(newAddress)
-        setIsValidAddress(ethers.isAddress(newAddress))
-    }, [])
-
-    const handleAmountChange = useCallback(
-        newAmount => {
-            setAmount(newAmount)
-            validateAmount(newAmount)
-        },
-        [maxAmount],
-    )
-
-    const validateAmount = useCallback(
-        value => {
-            const numValue = parseFloat(value)
-            const maxValue = parseFloat(maxAmount)
-            if (value === '') {
-                setAmountError('')
-            } else if (isNaN(numValue) || numValue <= 0) {
-                setAmountError('Amount must be greater than 0')
-            } else if (numValue > maxValue) {
-                setAmountError(`Amount cannot exceed ${maxAmount}`)
-            } else {
-                setAmountError('')
-            }
-        },
-        [maxAmount],
-    )
-
-    const handleMyAddressClick = useCallback(() => {
-        const newAddress = wallet.address
-        setAddress(newAddress)
-        setIsValidAddress(ethers.isAddress(newAddress))
-    }, [wallet.address])
-    const appDispatch = useAppDispatch()
-    const handleMaxAmountClick = useCallback(() => {
-        setAmount(maxAmount)
-        validateAmount(maxAmount)
-    }, [maxAmount, validateAmount])
-
-    async function handleWithdraw() {
-        setLoading(true)
-        if (!token) {
-            await withDraw(address, ethers.parseEther(amount))
-            getBalanceOfAnAddress(contractCMAccountAddress)
-        } else {
-            await transferERC20(token.address, address, ethers.parseEther(amount))
-            await fetchTokenBalances()
-        }
-        setAmount('')
-        setConfirm(false)
-        setAddress('')
-        appDispatch(
-            updateNotificationStatus({
-                message: 'Withdrawal completed successfully!',
-                severity: 'success',
-            }),
-        )
-        setOpen(false)
-        setLoading(false)
-    }
-
-    useEffect(() => {
-        getBalanceOfAnAddress(contractCMAccountAddress)
-    }, [getBalanceOfAnAddress])
-    return (
-        <Box sx={{ display: 'flex', flexDirection: 'column', gap: '12px' }}>
-            <AddressInput
-                address={address}
-                onAddressChange={handleAddressChange}
-                onMyAddressClick={handleMyAddressClick}
-            />
-            <AmountInput
-                amount={amount}
-                onAmountChange={handleAmountChange}
-                onMaxAmountClick={handleMaxAmountClick}
-                maxAmount={maxAmount}
-            />
-            <FormControlLabel
-                sx={{
-                    margin: '0 0',
-                }}
-                label={
-                    <Typography variant="body2">
-                        I double-checked the address I am about to send to
-                    </Typography>
-                }
-                control={
-                    <Checkbox
-                        sx={{
-                            color: theme => theme.palette.secondary.main,
-                            '&.Mui-checked': {
-                                color: theme => theme.palette.secondary.main,
-                            },
-                            '&.MuiCheckbox-colorSecondary.Mui-checked': {
-                                color: theme => theme.palette.secondary.main,
-                            },
-                            p: '0 8px 0 0',
-                        }}
-                        checked={confirm}
-                        onChange={() => setConfirm(prev => !prev)}
-                    />
-                }
-            />
-            {address !== '' && !isValidAddress && (
-                <Alert variant="negative" content="Invalid C-Chain address" />
-            )}
-            {amountError && <Alert variant="negative" content={amountError} />}
-            <Button
-                disabled={!confirm || !isValidAddress || !!amountError || !!!amount}
-                variant="contained"
-                onClick={handleWithdraw}
-                sx={{
-                    width: 'fit-content',
-                    padding: '6px 12px',
-                    gap: '6px',
-                    borderRadius: '8px',
-                    border: '1px solid #475569',
-                    borderWidth: '1px',
-                    '&:hover': {
-                        borderWidth: '1px',
-                        boxShadow: 'none',
-                    },
-                    flex: '0 0 16%',
-                    display: 'flex',
-                    alignItems: 'center',
-                    justifyContent: 'center',
-                }}
-            >
-                {loading ? (
-                    <CircularProgress size={20} style={{ width: 20, height: 20 }} color="inherit" />
-                ) : (
-                    <Typography variant="caption">Transfer</Typography>
-                )}
-            </Button>
-        </Box>
-    )
-}
-
 const MyMessenger = () => {
     const { state, dispatch } = usePartnerConfigurationContext()
-    const [open, setOpen] = useState(false)
-    const [selectedToken, setSelectedToken] = useState(null)
-    const [isOffChainPaymentSupported, setIsOffChainPaymentSupported] = useState(false)
-    const [isCAMSupported, setCAMSupported] = useState(false)
-    const [isEditMode, setIsEditMode] = useState(false)
-    const [tempOffChainPaymentSupported, setTempOffChainPaymentSupported] = useState(false)
-    const [tempSupportedTokens, setTempSupportedTokens] = useState([])
-    const [supportedTokens, setSupportedTokens] = useState([])
-    const [tempCAMSupported, setTempCAMSupported] = useState(false)
-    const { balanceOfAnAddress, getBalanceOfAnAddress } = useWalletBalance()
-    const [isLoading, setIsLoading] = useState(false)
     const [bots, setBots] = useState([])
-    const [tokens, setTokens] = useState([])
-    const { contractCMAccountAddress, wallet, provider } = useSmartContract()
-    const {
-        getSupportedTokens,
-        getOffChainPaymentSupported,
-        setOffChainPaymentSupported,
-        addSupportedToken,
-        removeSupportedToken,
-        getListOfBots,
-    } = usePartnerConfig()
+    const { getBalanceOfAnAddress } = useWalletBalance()
+    const { contractCMAccountAddress, wallet } = useSmartContract()
+    const { getListOfBots, sftAddress, sftName, sftSymbol } = usePartnerConfig()
     const partner = useAppSelector(rootState => selectPartnerData(rootState, '', wallet.address))
     const activeNetwork = useAppSelector(getActiveNetwork)
     useEffect(() => {
@@ -360,122 +32,18 @@ const MyMessenger = () => {
     }, [activeNetwork, dispatch])
 
     const appDispatch = useAppDispatch()
-    const handleOpenModal = token => {
-        setOpen(true)
-    }
-    async function checkIfOffChainPaymentSupported() {
-        let res = await getOffChainPaymentSupported()
-        setIsOffChainPaymentSupported(res)
-    }
 
-    const handleCloseModal = () => {
-        setSelectedToken(null)
-        setOpen(false)
-    }
-
-    const handleEditClick = () => {
-        const initialTempTokens = tokens.map(token => ({
-            ...token,
-            supported: supportedTokens.includes(token.address),
-        }))
-        setTempSupportedTokens(initialTempTokens)
-        setTempOffChainPaymentSupported(isOffChainPaymentSupported)
-        setTempCAMSupported(isCAMSupported)
-        setIsEditMode(true)
-    }
-
-    const handleCancelEdit = () => {
-        setIsEditMode(false)
-    }
-
-    const handleConfirmEdit = async () => {
-        setIsLoading(true)
-        try {
-            if (tempOffChainPaymentSupported !== isOffChainPaymentSupported) {
-                await setOffChainPaymentSupported(tempOffChainPaymentSupported)
-            }
-            if (tempCAMSupported !== isCAMSupported) {
-                if (tempCAMSupported) await addSupportedToken(ethers.ZeroAddress)
-                else await removeSupportedToken(ethers.ZeroAddress)
-            }
-            if (tempSupportedTokens) {
-                const excludeSet = new Set(supportedTokens)
-
-                for (const item of tempSupportedTokens) {
-                    const shouldBeSupported = !excludeSet.has(item.address)
-
-                    try {
-                        if (shouldBeSupported && item.supported) {
-                            await addSupportedToken(item.address)
-                        } else if (!shouldBeSupported && !item.supported) {
-                            await removeSupportedToken(item.address)
-                        }
-                    } catch (error) {
-                        console.error(`Error updating token ${item.address}:`, error)
-                    }
-                }
-            }
-            appDispatch(
-                updateNotificationStatus({
-                    message: 'Accepted currencies updated successfully',
-                    severity: 'success',
-                }),
-            )
-            setIsEditMode(false)
-        } catch (error) {
-            console.error('Error: ', error)
-        } finally {
-            await checkIfOffChainPaymentSupported()
-            await fetchSupportedTokens()
-            setIsLoading(false)
-        }
-    }
     async function fetchBots() {
         const res = await getListOfBots()
         setBots(res)
     }
-    async function fetchSupportedTokens() {
-        const res = await getSupportedTokens()
-        setCAMSupported(!!res.find(elem => elem === ethers.ZeroAddress))
-        setSupportedTokens(res)
-    }
-
-    const fetchTokenBalances = async () => {
-        const networkErc20Tokens = store.getters['Assets/networkErc20Tokens'] || []
-        if (networkErc20Tokens.length > 0) {
-            const fetchedTokens = await Promise.all(
-                networkErc20Tokens.map(async elem => {
-                    const contract = new ethers.Contract(
-                        elem.contract._address,
-                        ERC20_BALANCE_ABI,
-                        provider,
-                    )
-                    const balance = await contract.balanceOf(contractCMAccountAddress)
-                    return {
-                        address: elem.contract._address,
-                        balance: ethers.formatUnits(balance, elem.data.decimal),
-                        name: elem.data.name,
-                        symbol: elem.data.symbol,
-                        decimal: elem.data.decimal,
-                        supported: supportedTokens.includes(elem.contract._address),
-                    }
-                }),
-            )
-            setTokens(fetchedTokens)
-        }
-    }
-    useEffect(() => {
-        fetchTokenBalances()
-    }, [contractCMAccountAddress, supportedTokens])
 
     useEffect(() => {
         getBalanceOfAnAddress(contractCMAccountAddress)
     }, [contractCMAccountAddress])
 
     useEffect(() => {
-        checkIfOffChainPaymentSupported()
         fetchBots()
-        fetchSupportedTokens()
     }, [])
 
     function getServicesNames(services) {
@@ -505,8 +73,7 @@ const MyMessenger = () => {
                         {partner?.attributes?.companyName} Messenger Account
                     </Configuration.Title>
                     <Configuration.Paragraphe>
-                        In this page you are able to display and copy your Camino Messenger address,
-                        and manage accepted currencies.
+                        In this page you are able to display and copy your Camino Messenger address.
                     </Configuration.Paragraphe>
                     <TextField
                         disabled
@@ -603,7 +170,6 @@ const MyMessenger = () => {
                             )}
                         </Typography>
                     </Box>
-
                     {/* <ServiceList
                         listName="Wanted Services"
                         services={state.stepsConfig[2]?.services.map(elem => elem.name)}
@@ -642,394 +208,39 @@ const MyMessenger = () => {
                             )}
                         </Typography>
                     </Box>
-                    <Box
-                        sx={{
-                            display: 'flex',
-                            flexDirection: 'column',
-                            gap: '12px',
-                            width: 'fit-content',
-                        }}
-                    >
-                        <Box
-                            sx={{
-                                display: 'flex',
-                                justifyContent: 'space-between',
-                                alignItems: 'center',
-                            }}
-                        >
-                            <Typography variant="body2">Accepted Currencies</Typography>
-                            <RefreshOutlined
-                                onClick={() => {
-                                    getBalanceOfAnAddress(contractCMAccountAddress)
-                                    fetchTokenBalances()
-                                }}
-                                sx={{
-                                    cursor: 'pointer',
-                                    color: theme => `${theme.palette.text.primary} !important`,
-                                }}
-                            />
+                    <Box sx={{ display: 'flex', alignItems: 'center', gap: '16px' }}>
+                        <Typography sx={{ flex: '0 0 20%' }} variant="body2">
+                            Blockchain Transaction Fee Currency
+                        </Typography>
+                        <Box sx={{ display: 'flex', alignItems: 'center', gap: '8px' }}>
+                            <Typography variant="caption">CAM</Typography>
                         </Box>
-                        <FormControlLabel
-                            disabled={!isEditMode}
-                            label={<Typography variant="body2">Fiat: off-chain</Typography>}
-                            control={
-                                <Checkbox
-                                    sx={{
-                                        m: '0 8px 0 0',
-                                        color: theme =>
-                                            !isEditMode
-                                                ? theme.palette.action.disabled
-                                                : theme.palette.secondary.main,
-                                        '&.Mui-checked': {
-                                            color: theme =>
-                                                !isEditMode
-                                                    ? theme.palette.action.disabled
-                                                    : theme.palette.secondary.main,
-                                        },
-                                        '&.MuiCheckbox-colorSecondary.Mui-checked': {
-                                            color: theme =>
-                                                !isEditMode
-                                                    ? theme.palette.action.disabled
-                                                    : theme.palette.secondary.main,
-                                        },
+                    </Box>
+
+                    {/* Messenger fee currency */}
+                    <Box sx={{ display: 'flex', alignItems: 'center', gap: '16px' }}>
+                        <Typography sx={{ flex: '0 0 20%' }} variant="body2">
+                            Messenger Fee Currency
+                        </Typography>
+                        {sftAddress ? (
+                            <Box sx={{ display: 'flex', alignItems: 'center', gap: '8px' }}>
+                                <Blockies
+                                    seed={sftAddress.toLowerCase()}
+                                    size={8}
+                                    scale={3}
+                                    style={{
+                                        borderRadius: 8,
+                                        width: 24,
+                                        height: 24,
                                     }}
-                                    checked={
-                                        isEditMode
-                                            ? tempOffChainPaymentSupported
-                                            : isOffChainPaymentSupported
-                                    }
-                                    onChange={e =>
-                                        setTempOffChainPaymentSupported(e.target.checked)
-                                    }
                                 />
-                            }
-                        />
-                        <Box
-                            sx={{
-                                display: 'flex',
-                                alignItems: 'center',
-                                gap: '8px',
-                                justifyContent: 'space-between',
-                            }}
-                        >
-                            <FormControlLabel
-                                disabled={!isEditMode}
-                                label={
-                                    <Typography variant="body2">
-                                        CAM: {balanceOfAnAddress}
-                                    </Typography>
-                                }
-                                control={
-                                    <Checkbox
-                                        sx={{
-                                            m: '0 8px 0 0',
-                                            color: theme =>
-                                                !isEditMode
-                                                    ? theme.palette.action.disabled
-                                                    : theme.palette.secondary.main,
-                                            '&.Mui-checked': {
-                                                color: theme =>
-                                                    !isEditMode
-                                                        ? theme.palette.action.disabled
-                                                        : theme.palette.secondary.main,
-                                            },
-                                            '&.MuiCheckbox-colorSecondary.Mui-checked': {
-                                                color: theme =>
-                                                    !isEditMode
-                                                        ? theme.palette.action.disabled
-                                                        : theme.palette.secondary.main,
-                                            },
-                                        }}
-                                        checked={isEditMode ? tempCAMSupported : isCAMSupported}
-                                        onChange={e => setTempCAMSupported(e.target.checked)}
-                                    />
-                                }
-                            />
-                            {!isEditMode && (
-                                <Button
-                                    variant="contained"
-                                    onClick={() => {
-                                        setSelectedToken(null)
-                                        handleOpenModal({
-                                            symbol: 'CAM',
-                                        })
-                                    }}
-                                >
-                                    Withdraw
-                                </Button>
-                            )}
-                        </Box>
-                        {tokens.length > 0 &&
-                            tokens.map((elem, index) => {
-                                return (
-                                    <Box
-                                        sx={{
-                                            display: 'flex',
-                                            alignItems: 'center',
-                                            gap: '8px',
-                                            justifyContent: 'space-between',
-                                        }}
-                                        key={index}
-                                    >
-                                        <FormControlLabel
-                                            disabled={!isEditMode}
-                                            label={
-                                                <Typography variant="body2">
-                                                    {elem.name}: {elem.balance} {elem.symbol}
-                                                </Typography>
-                                            }
-                                            control={
-                                                <Checkbox
-                                                    sx={{
-                                                        m: '0 8px 0 0',
-                                                        color: theme =>
-                                                            !isEditMode
-                                                                ? theme.palette.action.disabled
-                                                                : theme.palette.secondary.main,
-                                                        '&.Mui-checked': {
-                                                            color: theme =>
-                                                                !isEditMode
-                                                                    ? theme.palette.action.disabled
-                                                                    : theme.palette.secondary.main,
-                                                        },
-                                                        '&.MuiCheckbox-colorSecondary.Mui-checked':
-                                                            {
-                                                                color: theme =>
-                                                                    !isEditMode
-                                                                        ? theme.palette.action
-                                                                              .disabled
-                                                                        : theme.palette.secondary
-                                                                              .main,
-                                                            },
-                                                    }}
-                                                    checked={
-                                                        isEditMode && tempSupportedTokens
-                                                            ? tempSupportedTokens[index]?.supported
-                                                            : elem?.supported
-                                                    }
-                                                    onChange={e => {
-                                                        let newArray = [...tempSupportedTokens]
-                                                        newArray[index].supported = e.target.checked
-                                                        setTempSupportedTokens(newArray)
-                                                    }}
-                                                />
-                                            }
-                                        />
-                                        {!isEditMode && (
-                                            <Button
-                                                variant="contained"
-                                                onClick={() => {
-                                                    setSelectedToken(elem)
-                                                    handleOpenModal(elem)
-                                                }}
-                                            >
-                                                Withdraw
-                                            </Button>
-                                        )}
-                                    </Box>
-                                )
-                            })}
-                        <Box
-                            sx={{
-                                display: 'flex',
-                                alignItems: 'center',
-                                gap: '8px',
-                                justifyContent: 'space-between',
-                            }}
-                        >
-                            <FormControlLabel
-                                label={
-                                    <Typography variant="body2">USDC* (coming soon) </Typography>
-                                }
-                                disabled
-                                control={
-                                    <Checkbox
-                                        sx={{
-                                            m: '0 8px 0 0',
-                                            color: theme =>
-                                                true
-                                                    ? theme.palette.action.disabled
-                                                    : theme.palette.secondary.main,
-                                            '&.Mui-checked': {
-                                                color: theme =>
-                                                    true
-                                                        ? theme.palette.action.disabled
-                                                        : theme.palette.secondary.main,
-                                            },
-                                            '&.MuiCheckbox-colorSecondary.Mui-checked': {
-                                                color: theme =>
-                                                    true
-                                                        ? theme.palette.action.disabled
-                                                        : theme.palette.secondary.main,
-                                            },
-                                        }}
-                                        checked={false}
-                                    />
-                                }
-                            />
-                            {!isEditMode && (
-                                <Button disabled={true} variant="contained">
-                                    Withdraw
-                                </Button>
-                            )}
-                        </Box>
-                        <Box
-                            sx={{
-                                display: 'flex',
-                                alignItems: 'center',
-                                gap: '8px',
-                                justifyContent: 'space-between',
-                            }}
-                        >
-                            <FormControlLabel
-                                disabled
-                                label={<Typography variant="body2">EURC* (coming soon)</Typography>}
-                                control={
-                                    <Checkbox
-                                        sx={{
-                                            m: '0 8px 0 0',
-                                            color: theme =>
-                                                true
-                                                    ? theme.palette.action.disabled
-                                                    : theme.palette.secondary.main,
-                                            '&.Mui-checked': {
-                                                color: theme =>
-                                                    true
-                                                        ? theme.palette.action.disabled
-                                                        : theme.palette.secondary.main,
-                                            },
-                                            '&.MuiCheckbox-colorSecondary.Mui-checked': {
-                                                color: theme =>
-                                                    true
-                                                        ? theme.palette.action.disabled
-                                                        : theme.palette.secondary.main,
-                                            },
-                                        }}
-                                        checked={false}
-                                    />
-                                }
-                            />
-                            {!isEditMode && (
-                                <Button disabled={true} variant="contained">
-                                    Withdraw
-                                </Button>
-                            )}
-                        </Box>
-                        <Box
-                            sx={{
-                                display: 'flex',
-                                alignItems: 'center',
-                                gap: '8px',
-                                justifyContent: 'space-between',
-                            }}
-                        >
-                            <FormControlLabel
-                                disabled
-                                label={<Typography variant="body2">EURe* (coming soon)</Typography>}
-                                control={
-                                    <Checkbox
-                                        sx={{
-                                            m: '0 8px 0 0',
-                                            color: theme =>
-                                                true
-                                                    ? theme.palette.action.disabled
-                                                    : theme.palette.secondary.main,
-                                            '&.Mui-checked': {
-                                                color: theme =>
-                                                    true
-                                                        ? theme.palette.action.disabled
-                                                        : theme.palette.secondary.main,
-                                            },
-                                            '&.MuiCheckbox-colorSecondary.Mui-checked': {
-                                                color: theme =>
-                                                    true
-                                                        ? theme.palette.action.disabled
-                                                        : theme.palette.secondary.main,
-                                            },
-                                        }}
-                                        checked={false}
-                                    />
-                                }
-                            />
-                            {!isEditMode && (
-                                <Button disabled={true} variant="contained">
-                                    Withdraw
-                                </Button>
-                            )}
-                        </Box>
-                        <Box
-                            sx={{
-                                display: 'flex',
-                                alignItems: 'center',
-                                gap: '8px',
-                                justifyContent: 'space-between',
-                            }}
-                        >
-                            <FormControlLabel
-                                disabled
-                                label={
-                                    <Typography variant="body2">EURSH* (coming soon)</Typography>
-                                }
-                                control={
-                                    <Checkbox
-                                        sx={{
-                                            m: '0 8px 0 0',
-                                            color: theme =>
-                                                true
-                                                    ? theme.palette.action.disabled
-                                                    : theme.palette.secondary.main,
-                                            '&.Mui-checked': {
-                                                color: theme =>
-                                                    true
-                                                        ? theme.palette.action.disabled
-                                                        : theme.palette.secondary.main,
-                                            },
-                                            '&.MuiCheckbox-colorSecondary.Mui-checked': {
-                                                color: theme =>
-                                                    true
-                                                        ? theme.palette.action.disabled
-                                                        : theme.palette.secondary.main,
-                                            },
-                                        }}
-                                        checked={false}
-                                    />
-                                }
-                            />
-                            {!isEditMode && (
-                                <Button disabled={true} variant="contained">
-                                    Withdraw
-                                </Button>
-                            )}
-                        </Box>
-                        <Box sx={{ display: 'flex', justifyContent: 'flex-start' }}>
-                            {!isEditMode ? (
-                                <Button variant="contained" onClick={handleEditClick}>
-                                    Configure Currencies
-                                </Button>
-                            ) : (
-                                <>
-                                    <Button
-                                        disabled={isLoading}
-                                        variant="outlined"
-                                        onClick={handleCancelEdit}
-                                        sx={{ mr: '8px' }}
-                                    >
-                                        Cancel
-                                    </Button>
-                                    <Button
-                                        disabled={isLoading}
-                                        variant="contained"
-                                        onClick={handleConfirmEdit}
-                                    >
-                                        {isLoading ? (
-                                            <CircularProgress size={24} color="inherit" />
-                                        ) : (
-                                            'Save Changes'
-                                        )}
-                                    </Button>
-                                </>
-                            )}
-                        </Box>
+                                <Typography variant="caption">
+                                    {sftName} ({sftSymbol})
+                                </Typography>
+                            </Box>
+                        ) : (
+                            <Typography variant="caption">Loading...</Typography>
+                        )}
                     </Box>
                 </Configuration>
                 <Box sx={{ display: 'flex', flexDirection: 'column', gap: '16px' }}>
@@ -1043,48 +254,6 @@ const MyMessenger = () => {
                     ></Configuration.Infos>
                 </Box>
             </Box>
-            <DialogAnimate open={open} onClose={handleCloseModal}>
-                <DialogTitle
-                    sx={{
-                        m: 0,
-                        p: 2,
-                        width: '768px',
-                        backgroundColor: theme =>
-                            theme.palette.mode === 'dark' ? '#020617' : '#F1F5F9',
-                    }}
-                >
-                    <Typography variant="body1" component="span">
-                        Withdraw {selectedToken ? selectedToken.symbol : 'CAM'}
-                    </Typography>
-                    <IconButton
-                        aria-label="close"
-                        onClick={handleCloseModal}
-                        sx={{
-                            position: 'absolute',
-                            right: 10,
-                            top: 15,
-                            cursor: 'pointer',
-                            color: theme => theme.palette.grey[500],
-                        }}
-                    >
-                        <Icon path={mdiClose} size={1} />
-                    </IconButton>
-                </DialogTitle>
-
-                <Divider sx={{ borderWidth: '1.5px' }} />
-                <DialogContent
-                    sx={{
-                        backgroundColor: theme =>
-                            theme.palette.mode === 'dark' ? '#020617' : '#F1F5F9',
-                    }}
-                >
-                    <CamWithdraw
-                        setOpen={setOpen}
-                        token={selectedToken}
-                        fetchTokenBalances={fetchTokenBalances}
-                    />
-                </DialogContent>
-            </DialogAnimate>
         </>
     )
 }
